+++ custom_components/octopus_energy_de/api/graphql/queries.py
@@ -135,3 +135,57 @@ query ElectricityConsumption($accountNumber: String!, $propertyId: ID!, $date: D
   }
 }
 """
+
+SMARTFLEX_QUERY = """
+query SmartFlex($accountNumber: String!) {
+  completedDispatches(accountNumber: $accountNumber) {
+    deltaKwh
+    start
+    startDt
+    end
+    endDt
+  }
+  devices(accountNumber: $accountNumber) {
+    id
+    name
+    deviceType
+    provider
+    status {
+      currentState
+      isSuspended
+      ... on SmartFlexVehicleStatus {
+        activePower { value }
+        stateOfCharge { value }
+      }
+      ... on SmartFlexChargePointStatus {
+        stateOfCharge { value }
+      }
+    }
+    ... on SmartFlexVehicle {
+      vehicleVariant { batterySize }
+    }
+    ... on SmartFlexVehicle {
+      chargingSessions(first: 100) {
+        edges { node {
+          start
+          end
+          energyAdded { value }
+          cost { amount }
+          ... on SmartFlexChargingSession { type }
+        }}
+      }
+    }
+    ... on SmartFlexChargePoint {
+      chargingSessions(first: 100) {
+        edges { node {
+          start
+          end
+          energyAdded { value }
+          cost { amount }
+          ... on SmartFlexChargingSession { type }
+        }}
+      }
+    }
+  }
+}
+"""